// src/enums.ts
export enum Sorting {
  None = 'None',
  PriceAsc = 'Price (asc)',
  PriceDesc = 'Price (desc)',
}

export enum WineType {
  Red = 'Red',
  White = 'White',
  Sparkling = 'Sparkling',
  Rose = 'Rosé',
}

export enum Country {
  Argentina = 'Argentina',
  Australia = 'Australia',
  Chile = 'Chile',
  France = 'France',
  Georgia = 'Georgia',
  Germany = 'Germany',
  Greece = 'Greece',
  Italy = 'Italy',
  NewZealand = 'New Zealand',
  Portugal = 'Portugal',
  SouthAfrica = 'South Africa',
  Spain = 'Spain',
  UK = 'United Kingdom',
}